Added stages logic into N&S, so that parts of it can be run in isolation

+++ lib/forklift/__init__.py
@@ -10,7 +10,7 @@ from .stages import *
 
 class Forklift(object):
   @classmethod
-  def normalize_and_sanitize(klass, dataframe, with_spark_schema, remappings_file_path, cast_processor):
+  def normalize_and_sanitize(klass, dataframe, with_spark_schema, remappings_file_path, cast_processor, stages=NS_ALL):
     print("Step 1: Check that arguments are valid")
     if not isinstance(dataframe, DataFrame):
       raise TypeError("dataframe must be an instance of Dataframe")
@@ -26,18 +26,29 @@ class Forklift(object):
       raise ValueError("remappings_file_path must be a file that exists")
     if not isinstance(cast_processor, CastProcessor):
       raise TypeError("cast_processor must be an instance of CastProcessor")
+    if not isinstance(stages, ForkliftStage):
+      raise TypeError("stages must be an instance of ForkliftStage")
 
-    print("Step 2: Rename all columns, according to the mapping")
-    column_renamer = Forklift.ColumnRenamer(remappings_file_path)
-    dataframe_with_columns_renamed = column_renamer.rename_columns(dataframe)
+    if any([stage in stages for stage in [NS_ALL, NS_RENAME_COLS]]):
+      print("Step 2: Rename all columns, according to the mapping")
+      column_renamer = Forklift.ColumnRenamer(remappings_file_path)
+      dataframe_with_columns_renamed = column_renamer.rename_columns(dataframe)
+    else:
+      print("Skipping Step 2: Rename all columns, according to the mapping")
 
-    print("Step 3: Delete unwanted columns, according to the mapping")
-    column_deleter = Forklift.ColumnDeleter(remappings_file_path)
-    dataframe_with_columns_renamed_and_unwanted_columns_deleted = column_deleter.delete_columns(dataframe_with_columns_renamed)
+    if any([stage in stages for stage in [NS_ALL, NS_DELETE_COLS]]):
+      print("Step 3: Delete unwanted columns, according to the mapping")
+      column_deleter = Forklift.ColumnDeleter(remappings_file_path)
+      dataframe_with_columns_renamed_and_unwanted_columns_deleted = column_deleter.delete_columns(dataframe_with_columns_renamed)
+    else:
+      print("Skipping Step 3: Delete unwanted columns, according to the mapping")
 
-    print("Step 4: Cast each cell, according to the Caster instance provided")
-    caster = Forklift.Caster(cast_processor, with_spark_schema)
-    dataframe_with_columns_renamed_and_unwanted_columns_deleted_and_types_cast_and_schema_applied = caster.cast(dataframe_with_columns_renamed_and_unwanted_columns_deleted)
+    if any([stage in stages for stage in [NS_ALL, NS_CAST_CELLS]]):
+      print("Step 4: Cast each cell, according to the Caster instance provided")
+      caster = Forklift.Caster(cast_processor, with_spark_schema)
+      dataframe_with_columns_renamed_and_unwanted_columns_deleted_and_types_cast_and_schema_applied = caster.cast(dataframe_with_columns_renamed_and_unwanted_columns_deleted)
+    else:
+      print("Skipping Step 4: Cast each cell, according to the Caster instance provided")
 
     print("Step 5: Return the normalized and sanitized dataframe")
     return dataframe_with_columns_renamed_and_unwanted_columns_deleted_and_types_cast_and_schema_applied
